Mirrored from local repository @ 01/17/2022 22:14

+++ Source/Control/UI/EditControlDerives/HtmlEditControl.Main.cs
@@ -40,11 +40,7 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
 
             if (!DesignMode && !HtmlEditorDesignModeManager.IsDesignMode)
             {
-                AllowWebBrowserDrop = false;
-
-                Navigate(@"about:blank");
-
-                _tmpFolderPath = Path.Combine(Path.GetTempPath(), @"zhe1-"+ Guid.NewGuid());
+                _tmpFolderPath = Path.Combine(Path.GetTempPath(), @"zhe1-" + Guid.NewGuid());
                 Directory.CreateDirectory(_tmpFolderPath);
 
                 _timerTextChange.Tick += timerTextChange_Tick;
@@ -57,6 +53,26 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
 
                 Configure(Configuration);
             }
+        }
+
+        // Siehe andere "Bugfix 2015-11-02"-Kommentare.
+        [Browsable(false)]
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
+        public new bool AllowWebBrowserDrop
+        {
+            get { return base.AllowWebBrowserDrop; }
+            set { base.AllowWebBrowserDrop = value; }
+        }
+
+        // Siehe andere "Bugfix 2015-11-02"-Kommentare.
+        [Browsable(false)]
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
+        public new bool IsWebBrowserContextMenuEnabled
+        {
+            get { return base.IsWebBrowserContextMenuEnabled; }
+            set { base.IsWebBrowserContextMenuEnabled = value; }
         }
 
         [Browsable(false)]
@@ -83,10 +99,51 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
             }
         }
 
+        protected override void OnCheckInitialize(EventArgs args)
+        {
+            base.OnCheckInitialize(args);
+
+            // --
+
+            if (!DesignMode && !HtmlEditorDesignModeManager.IsDesignMode && !EverInitialized && IsHandleCreated)
+            {
+                var form = FindForm();
+                if (form != null)
+                {
+                    // Uwe Keim: Bugfix 2015-11-02:
+                    // 
+                    // Wenn auf Document-Instanz zu früh zu gegriffen wird, hat das den Effekt,
+                    // dass bei mehr als einem Editor auf einem Form, das Caret in allen
+                    // Editoren blinkt.
+                    //
+                    // Deshalb habe ich hierher einige Funktionsaufrufe gesetzt, die vorher im
+                    // Konstruktor standen.
+                    //
+                    // Das alleine hat es noch nicht ausgemacht, geholfen hat auch:
+                    //
+                    // Mehrfach den Fokus zwischen den einzelnen HTML-Editoren ändern,
+                    // lässt den Fehler verschwinden, dass mehrere Carets gleichzeitig blinken.
+
+                    if (!DesignMode && !HtmlEditorDesignModeManager.IsDesignMode)
+                    {
+                        AllowWebBrowserDrop = false;
+
+                        Navigate(@"about:blank");
+                    }
+
+                    // --
+
+                    MarkAsEverInitialized();
+                }
+            }
+        }
+
         protected override void OnHandleDestroyed(EventArgs e)
         {
             if (!DesignMode && !HtmlEditorDesignModeManager.IsDesignMode)
             {
+                UnmarkAsEverInitialized();
+
                 if (_timerTextChange != null)
                 {
                     _timerTextChange.Stop();
@@ -96,10 +153,7 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
 
                 if (!string.IsNullOrEmpty(_tmpFolderPath))
                 {
-                    if (Directory.Exists(_tmpFolderPath))
-                    {
-                        Directory.Delete(_tmpFolderPath, true);
-                    }
+                    if (Directory.Exists(_tmpFolderPath)) Directory.Delete(_tmpFolderPath, true);
                     _tmpFolderPath = null;
                 }
             }
@@ -115,19 +169,15 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
             updateUI();
         }
 
+        public new HtmlDocument Document => EverInitialized ? base.Document : null;
+
         protected override void OnGotFocus(EventArgs e)
         {
             base.OnGotFocus(e);
 
-            if (!DesignMode && !HtmlEditorDesignModeManager.IsDesignMode)
+            if (!DesignMode && !HtmlEditorDesignModeManager.IsDesignMode && EverInitialized)
             {
-                if (Document != null)
-                {
-                    if (Document.Body != null)
-                    {
-                        Document.Body.Focus();
-                    }
-                }
+                Document?.Body?.Focus();
             }
         }
 
@@ -151,8 +201,7 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
             object sender,
             EventArgs e)
         {
-            if (!DesignMode && !HtmlEditorDesignModeManager.IsDesignMode) {
-                if (!IsDisposed) // Uwe Keim 2006-03-17.
+            if (!DesignMode && !HtmlEditorDesignModeManager.IsDesignMode && !IsDisposed && IsHandleCreated)
             {
                 var s = DocumentText ?? string.Empty;
 
@@ -161,9 +210,9 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
                     _tmpCacheTextChange = s;
 
                     var h = TextChanged;
-                    if (h != null) h(this, new EventArgs());
+                    h?.Invoke(this, new EventArgs());
                 }
-            }}
+            }
         }
 
         public new event EventHandler TextChanged;
@@ -175,7 +224,9 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
 
         private void updateUI()
         {
-            if (Document != null && Document.DomDocument != null)
+            // Siehe andere "Bugfix 2015-11-02"-Kommentare.
+
+            if ( EverInitialized && Document?.DomDocument != null)
             {
                 boldToolStripMenuItem.Enabled = CanBold;
                 italicToolStripMenuItem.Enabled = CanItalic;
@@ -223,166 +274,199 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
 
         private void boldToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteBold();
         }
 
         private void italicToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteItalic();
         }
 
         private void pasteToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecutePaste();
         }
 
         private void pasteAsTextToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecutePasteAsText();
         }
 
         private void htmlToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteShowSource();
         }
 
         private void hyperLinkToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteInsertHyperlink();
         }
 
         private void indentToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteIndent();
         }
 
         private void justifyCenterToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteJustifyCenter();
         }
 
         private void justifyLeftToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteJustifyLeft();
         }
 
         private void justifyRightToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteJustifyRight();
         }
 
         private void numberedListToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteNumberedList();
         }
 
         private void outdentToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteOutdent();
         }
 
         private void bullettedListToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteBullettedList();
         }
 
         private void copyToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteCopy();
         }
 
         private void cutToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteCut();
         }
 
         private void deleteToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteDelete();
         }
 
         private void foreColorNoneToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteSetForeColorNone();
         }
 
         private void foreColor01ToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteSetForeColor01();
         }
 
         private void foreColor02ToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteSetForeColor02();
         }
 
         private void foreColor03ToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteSetForeColor03();
         }
 
         private void foreColor04ToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteSetForeColor04();
         }
 
         private void foreColor05ToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteSetForeColor05();
         }
 
         private void foreColor06ToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteSetForeColor06();
         }
 
         private void foreColor07ToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteSetForeColor07();
         }
 
         private void foreColor08ToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteSetForeColor08();
         }
 
         private void foreColor09ToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteSetForeColor09();
         }
 
         private void foreColor10ToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteSetForeColor10();
         }
 
         private void backColorNoneToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteSetBackColorNone();
         }
 
         private void backColor01ToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteSetBackColor01();
         }
 
         private void backColor02ToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteSetBackColor02();
         }
 
         private void backColor03ToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteSetBackColor03();
         }
 
         private void backColor04ToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteSetBackColor04();
         }
 
         private void backColor05ToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteSetBackColor05();
         }
 
@@ -390,6 +474,7 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
             object sender,
             EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteInsertTable();
         }
 
@@ -397,6 +482,7 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
             object sender,
             EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteInsertTableRow();
         }
 
@@ -404,6 +490,7 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
             object sender,
             EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteInsertTableColumn();
         }
 
@@ -411,6 +498,7 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
             object sender,
             EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteTableAddTableRow();
         }
 
@@ -418,6 +506,7 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
             object sender,
             EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteTableAddTableColumn();
         }
 
@@ -425,6 +514,7 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
             object sender,
             EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteTableProperties();
         }
 
@@ -432,6 +522,7 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
             object sender,
             EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteTableRowProperties();
         }
 
@@ -439,6 +530,7 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
             object sender,
             EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteTableColumnProperties();
         }
 
@@ -446,6 +538,7 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
             object sender,
             EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteTableCellProperties();
         }
 
@@ -453,6 +546,7 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
             object sender,
             EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteTableDeleteRow();
         }
 
@@ -460,6 +554,7 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
             object sender,
             EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteTableDeleteColumn();
         }
 
@@ -467,15 +562,17 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
             object sender,
             EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteTableDeleteTable();
         }
 
         protected override void OnUpdateUI()
         {
+            if (!EverInitialized) return;
             if (!DesignMode && !HtmlEditorDesignModeManager.IsDesignMode)
             {
                 var h = UINeedsUpdate;
-                if (h != null) h(this, EventArgs.Empty);
+                h?.Invoke(this, EventArgs.Empty);
             }
         }
 
@@ -504,7 +601,8 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
 
         private void checkGetTextModules()
         {
-            if (Configuration != null && Configuration.ExternalInformationProvider != null && !_everLoadedTextModules)
+            if (!EverInitialized) return;
+            if (Configuration?.ExternalInformationProvider != null && !_everLoadedTextModules)
             {
                 _everLoadedTextModules = true;
                 _textModules = Configuration.ExternalInformationProvider.GetTextModules();
@@ -513,11 +611,13 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
 
         private void textModulesToolStripMenuItem_DropDownOpening(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             checkFillTextModules(textModulesToolStripMenuItem);
         }
 
         private void checkFillTextModules(ToolStripDropDownItem toolStripMenuItem)
         {
+            if (!EverInitialized) return;
             if (!_textModulesFilled)
             {
                 _textModulesFilled = true;
@@ -531,11 +631,13 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
             NativeMethods.IUnknown queryForStatus,
             NativeMethods.IDispatch objectAtScreenCoordinates)
         {
+            //if (!EverInitialized) return true;
+
             base.OnNeedShowContextMenu(contextMenuKind, position, queryForStatus, objectAtScreenCoordinates);
 
             if (!DesignMode && !HtmlEditorDesignModeManager.IsDesignMode)
             {
-                if (Configuration != null && Configuration.ExternalInformationProvider != null)
+                if (Configuration?.ExternalInformationProvider != null)
                 {
                     var font = Configuration.ExternalInformationProvider.Font;
                     contextMenuStrip.Font = font ?? Font;
@@ -558,16 +660,19 @@ namespace ZetaHtmlEditControl.UI.EditControlDerives
 
         private void underlineToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteUnderline();
         }
 
         private void pasteFromMsWordToolStripItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecutePasteFromWord();
         }
 
         private void removeFormattingToolStripMenuItem_Click(object sender, EventArgs e)
         {
+            if (!EverInitialized) return;
             ExecuteRemoveFormatting();
         }
     }
